fix: zotero download logic in the utils
